[Pal/Linux-SGX] Fix stack unwinding on enclave exit
A previous change in LibOS started storing fake syscall return
address in R8. This broke stack unwinding when the stack spans both
LibOS syscall and SGX OCALL, because we had no CFI directives for
recovering register values except for RBP, RSP and RIP, and the
value of R8 was lost.

This change adds proper CFI to enclave exit code, which is
unfortunately pretty tricky due to switching stacks. As a result
GDB should have access to all saved registers.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>